feat(context): revert back traditional react context
